refactor(core): split service module into install and message modules
